fix: icon not found
Some users may not have icons and the icon plugin is not a dependency
for vgit.nvim. In that instance, we do not compute any icon showing
logic and additionally there is a further check now to see if the plugin
I am targetting for icons is loaded or not.